porting to GCC... 9

+++ samples/slept/sleepnew.cpp
@@ -69,7 +69,7 @@ int __cdecl main(void)
     }
 #endif
 
-    printf("sleepnew.exe: GetSleptTicks() = %d\n\n", GetSleptTicks());
+    printf("sleepnew.exe: GetSleptTicks() = %d\n\n", (int)GetSleptTicks());
     return 0;
 }
 //
